Drop down menu.
I added the drop down menu. I know the drop down menu aspect of it
works. But since I can't see the courses I am not sure if it loads them
correctly or not. I'm also not sure if where it is located is a problem
with where the classes are loaded.